[Refactor] 文字コード関連処理のファイルを再構成
文字コード関連の処理が複数のファイルに散らばっているので、
src/locale/character-encoding.cpp,h にまとめる。
