+++ VisualStudio/Hengband/Hengband.vcxproj
@@ -312,6 +312,7 @@
     <ClCompile Include="..\..\src\io\macro-configurations-store.cpp" />
     <ClCompile Include="..\..\src\io\temp-file.cpp" />
     <ClCompile Include="..\..\src\load\artifact-record-loader.cpp" />
+    <ClCompile Include="..\..\src\locale\character-encoding.cpp" />
     <ClCompile Include="..\..\src\monster-floor\monster-movement-direction-list.cpp" />
     <ClCompile Include="..\..\src\player-info\bluemage-data.cpp" />
     <ClCompile Include="..\..\src\save\artifact-record-writer.cpp" />
@@ -492,7 +493,6 @@
     <ClCompile Include="..\..\src\io\input-key-requester.cpp" />
     <ClCompile Include="..\..\src\store\store-key-processor.cpp" />
     <ClCompile Include="..\..\src\load\info-loader.cpp" />
-    <ClCompile Include="..\..\src\locale\utf-8.cpp" />
     <ClCompile Include="..\..\src\main-win\main-win-cfg-reader.cpp" />
     <ClCompile Include="..\..\src\main-win\main-win-mci.cpp" />
     <ClCompile Include="..\..\src\main-win\main-win-music.cpp" />
@@ -1318,7 +1318,6 @@
     <ClInclude Include="..\..\src\load\info-loader.h" />
     <ClInclude Include="..\..\src\locale\character-encoding.h" />
     <ClInclude Include="..\..\src\locale\language-switcher.h" />
-    <ClInclude Include="..\..\src\locale\utf-8.h" />
     <ClInclude Include="..\..\src\main-win\main-win-cfg-reader.h" />
     <ClInclude Include="..\..\src\main-win\main-win-mci.h" />
     <ClInclude Include="..\..\src\main-win\main-win-menuitem.h" />
