+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -2021,9 +2021,6 @@
     <ClCompile Include="..\..\src\system\angband-version.cpp">
       <Filter>system</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\locale\utf-8.cpp">
-      <Filter>locale</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\wizard\wizard-game-modifier.cpp">
       <Filter>wizard</Filter>
     </ClCompile>
@@ -2606,6 +2603,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\load\artifact-record-loader.cpp">
       <Filter>load\item</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\locale\character-encoding.cpp">
+      <Filter>locale</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -4876,9 +4876,6 @@
     </ClInclude>
     <ClInclude Include="..\..\src\artifact\random-art-effects.h">
       <Filter>artifact</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\locale\utf-8.h">
-      <Filter>locale</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\wizard\wizard-game-modifier.h">
       <Filter>wizard</Filter>
