+++ src/birth/history-editor.cpp
@@ -1,7 +1,7 @@
 #include "birth/history-editor.h"
 #include "io/input-key-acceptor.h"
 #include "io/read-pref-file.h"
-#include "locale/japanese.h"
+#include "locale/character-encoding.h"
 #include "system/player-type-definition.h"
 #include "term/screen-processor.h"
 #include "term/term-color-types.h"
