+++ src/info-reader/json-reader-util.cpp
@@ -1,6 +1,6 @@
 #include "info-reader/json-reader-util.h"
 #include "info-reader/info-reader-util.h"
-#include "locale/japanese.h"
+#include "locale/character-encoding.h"
 #include "util/dice.h"
 
 /*!
