+++ src/locale/japanese.cpp
@@ -5,11 +5,7 @@
  */
 
 #include "locale/japanese.h"
-#include "locale/character-encoding.h"
-#include "locale/utf-8.h"
 #include "util/enum-converter.h"
-#include "util/string-processor.h"
-#include "view/display-messages.h"
 #include <set>
 #include <sstream>
 
@@ -168,450 +164,6 @@ static const std::set<std::string_view> kinsoku_list{
 bool is_kinsoku(std::string_view ch)
 {
     return (ch.length() >= 2) && kinsoku_list.contains(ch);
-}
-
-/*!
- * @brief 文字コードをSJISからEUCに変換する / Convert SJIS string to EUC string
- * @param str 変換する文字列のポインタ
- * @details
- */
-void sjis2euc(char *str)
-{
-    int i;
-    unsigned char c1, c2;
-
-    int len = strlen(str);
-
-    std::vector<char> tmp(len + 1);
-
-    for (i = 0; i < len; i++) {
-        c1 = str[i];
-        if (c1 & 0x80) {
-            i++;
-            c2 = str[i];
-            if (c2 >= 0x9f) {
-                c1 = c1 * 2 - (c1 >= 0xe0 ? 0xe0 : 0x60);
-                c2 += 2;
-            } else {
-                c1 = c1 * 2 - (c1 >= 0xe0 ? 0xe1 : 0x61);
-                c2 += 0x60 + (c2 < 0x7f);
-            }
-            tmp[i - 1] = c1;
-            tmp[i] = c2;
-        } else {
-            tmp[i] = c1;
-        }
-    }
-    tmp[len] = 0;
-    strcpy(str, tmp.data());
-}
-
-/*!
- * @brief 文字コードをEUCからSJISに変換する / Convert EUC string to SJIS string
- * @param str 変換する文字列のポインタ
- * @details
- */
-void euc2sjis(char *str)
-{
-    int i;
-    unsigned char c1, c2;
-
-    int len = strlen(str);
-
-    std::vector<char> tmp(len + 1);
-
-    for (i = 0; i < len; i++) {
-        c1 = str[i];
-        if (c1 & 0x80) {
-            i++;
-            c2 = str[i];
-            if (c1 % 2) {
-                c1 = (c1 >> 1) + (c1 < 0xdf ? 0x31 : 0x71);
-                c2 -= 0x60 + (c2 < 0xe0);
-            } else {
-                c1 = (c1 >> 1) + (c1 < 0xdf ? 0x30 : 0x70);
-                c2 -= 2;
-            }
-
-            tmp[i - 1] = c1;
-            tmp[i] = c2;
-        } else {
-            tmp[i] = c1;
-        }
-    }
-    tmp[len] = 0;
-    strcpy(str, tmp.data());
-}
-
-/*!
- * @brief strを環境に合った文字コードに変換し、変換前の文字コードを返す。strの長さに制限はない。
- * @param str 変換する文字列のポインタ
- * @return 変換前の文字コード
- */
-CharacterEncoding codeconv(char *str)
-{
-    auto encoding = CharacterEncoding::UNKNOWN;
-    for (auto i = 0; str[i]; i++) {
-        /* First byte */
-        const auto c1 = static_cast<unsigned char>(str[i]);
-
-        /* ASCII? */
-        if (!(c1 & 0x80)) {
-            encoding = CharacterEncoding::US_ASCII;
-            continue;
-        }
-
-        /* Second byte */
-        i++;
-        const auto c2 = static_cast<unsigned char>(str[i]);
-
-        const auto is_euc_jp = ((0xa1 <= c1 && c1 <= 0xdf) || (0xfd <= c1 && c1 <= 0xfe)) && (0xa1 <= c2 && c2 <= 0xfe);
-        auto is_cp932 = (0x81 <= c1 && c1 <= 0x9f) && ((0x40 <= c2 && c2 <= 0x7e) || (0x80 <= c2 && c2 <= 0xfc));
-        is_cp932 |= (0xe0 <= c1 && c1 <= 0xfc) && (0x40 <= c2 && c2 <= 0x7e);
-
-        if (is_euc_jp) {
-            /* Only EUC is allowed */
-            if (encoding == CharacterEncoding::UNKNOWN || encoding == CharacterEncoding::US_ASCII || encoding == CharacterEncoding::EUC_JP) {
-                encoding = CharacterEncoding::EUC_JP;
-                continue;
-            }
-        } else if (is_cp932) {
-            /* Only SJIS is allowed */
-            if (encoding == CharacterEncoding::UNKNOWN || encoding == CharacterEncoding::US_ASCII || encoding == CharacterEncoding::SHIFT_JIS) {
-                encoding = CharacterEncoding::SHIFT_JIS;
-                continue;
-            }
-        }
-
-        /* Broken string, no conversion */
-        return CharacterEncoding::UNKNOWN;
-    }
-
-    switch (encoding) {
-#ifdef EUC
-    case CharacterEncoding::SHIFT_JIS:
-        sjis2euc(str);
-        break;
-#endif
-
-#ifdef SJIS
-    case CharacterEncoding::EUC_JP:
-        euc2sjis(str);
-        break;
-#endif
-    default:
-        break;
-    }
-
-    return encoding;
-}
-
-/*!
- * @brief 文字列sのxバイト目が漢字の1バイト目かどうか判定する
- * @param s 判定する文字列のポインタ
- * @param x 判定する位置(バイト)
- * @return 漢字の1バイト目ならばTRUE
- */
-bool iskanji2(concptr s, int x)
-{
-    int i;
-
-    for (i = 0; i < x; i++) {
-        if (iskanji(s[i])) {
-            i++;
-        }
-    }
-    if ((x == i) && iskanji(s[x])) {
-        return true;
-    }
-
-    return false;
-}
-
-/*!
- * @brief 文字列の文字コードがASCIIかどうかを判定する
- * @param str 判定する文字列へのポインタ
- * @return 文字列の文字コードがASCIIならTRUE、そうでなければFALSE
- */
-static bool is_ascii_str(concptr str)
-{
-    for (; *str; str++) {
-        int ch = *str;
-        if (!(0x00 < ch && ch <= 0x7f)) {
-            return false;
-        }
-    }
-    return true;
-}
-
-#if defined(EUC)
-#include <algorithm>
-#include <iconv.h>
-#include <initializer_list>
-#include <vector>
-
-// UTF-8 の文字列長は必ずしも3バイトとは限らないが、変愚蛮怒の仕様範囲では3固定.
-constexpr auto ENCODING_LENGTH = 3;
-class EncodingConverter {
-public:
-    EncodingConverter(const std::initializer_list<unsigned char> from, const std::initializer_list<unsigned char> to)
-        : from(from)
-        , to(to)
-    {
-    }
-
-    bool equals(const unsigned char *p) const
-    {
-        return std::equal(from.begin(), from.end(), p);
-    }
-
-    void replace(unsigned char *p) const
-    {
-        std::copy_n(to.begin(), ENCODING_LENGTH, p);
-    }
-
-private:
-    std::vector<unsigned char> from;
-    std::vector<unsigned char> to;
-};
-
-const std::vector<EncodingConverter> encoding_characters = {
-    { { 0xef, 0xbd, 0x9e }, { 0xe3, 0x80, 0x9c } }, /* FULLWIDTH TILDE -> WAVE DASH (全角チルダ → 波ダッシュ) */
-    { { 0xef, 0xbc, 0x8d }, { 0xe2, 0x88, 0x92 } }, /* FULLWIDTH HYPHEN-MINUS -> MINUS SIGN (全角ハイフン → マイナス記号) */
-};
-
-/*!
- * @brief 受け取ったUTF-8文字列を調べ、特定のコードポイントの文字の置き換えを行う
- *
- * 受け取ったUTF-8の文字列に含まれる文字を1つ1つ調べて、encoding_characters で
- * 定義されている特定のコードポイントの文字の変換を行う。
- *
- * '～'と'－'は、Windows環境(CP932)とLinux/UNIX環境(EUC-JP)でUTF-8に対応する
- * 文字としてそれぞれ別のコードポイントが割り当てられており、別の環境の
- * UTF-8からシステムの文字コードに変換した時に、これらの文字は変換できず
- * 文字化けが起きてしまう。
- *
- * これを避けるため、Linux/UNIX環境(EUC-JP)ではUTF-8→EUC-JPの変換を行う前に
- * 該当するコードポイントの文字をLinux/UNIX環境のものに置き換えてから
- * 変換を行うようにするためのルーチン。
- *
- * @param str コードポイントの置き換えを行う文字列へのポインタ
- */
-static void ms_to_jis_unicode(char *str)
-{
-    for (auto *p = (unsigned char *)str; *p; p++) {
-        auto subseq_num = 0;
-        if (0x00 < *p && *p <= 0x7f) {
-            continue;
-        }
-
-        if ((*p & 0xe0) == 0xc0) {
-            subseq_num = 1;
-        }
-
-        if ((*p & 0xf0) == 0xe0) {
-            for (const auto &converter : encoding_characters) {
-                if (converter.equals(p)) {
-                    converter.replace(p);
-                }
-            }
-
-            subseq_num = 2;
-        }
-
-        if ((*p & 0xf8) == 0xf0) {
-            subseq_num = 3;
-        }
-
-        p += subseq_num;
-    }
-}
-
-#elif defined(SJIS) && defined(WINDOWS)
-#include <Windows.h>
-#endif
-
-#ifdef EUC
-/*!
- * @brief 文字列の文字コードをUTF-8からEUC-JPに変換する
- * @param utf8_str 変換元の文字列へのポインタ
- * @param utf8_str_len 変換元の文字列の長さ(文字数ではなくバイト数)
- * @param euc_buf 変換した文字列を格納するバッファへのポインタ
- * @param euc_buf_len 変換した文字列を格納するバッファのサイズ
- * @return 変換に成功した場合変換後の文字列の長さを返す
- *         変換に失敗した場合-1を返す
- */
-int utf8_to_euc(char *utf8_str, size_t utf8_str_len, char *euc_buf, size_t euc_buf_len)
-{
-    static iconv_t cd = nullptr;
-    if (!cd) {
-        cd = iconv_open("EUC-JP", "UTF-8");
-    }
-
-    ms_to_jis_unicode(utf8_str);
-
-    size_t inlen_left = utf8_str_len;
-    size_t outlen_left = euc_buf_len;
-    char *in = utf8_str;
-    char *out = euc_buf;
-
-    if (iconv(cd, &in, &inlen_left, &out, &outlen_left) == (size_t)-1) {
-        return -1;
-    }
-
-    return euc_buf_len - outlen_left;
-}
-
-/*!
- * @brief 文字列の文字コードをEUC-JPからUTF-8に変換する
- * @param euc_str 変換元の文字列へのポインタ
- * @param euc_str_len 変換元の文字列の長さ(文字数ではなくバイト数)
- * @param utf8_buf 変換した文字列を格納するバッファへのポインタ
- * @param utf8_buf_len 変換した文字列を格納するバッファのサイズ
- * @return 変換に成功した場合変換後の文字列の長さを返す
- *         変換に失敗した場合-1を返す
- */
-int euc_to_utf8(const char *euc_str, size_t euc_str_len, char *utf8_buf, size_t utf8_buf_len)
-{
-    static iconv_t cd = nullptr;
-    if (!cd) {
-        cd = iconv_open("UTF-8", "EUC-JP");
-    }
-
-    size_t inlen_left = euc_str_len;
-    size_t outlen_left = utf8_buf_len;
-    const char *in = euc_str;
-    char *out = utf8_buf;
-
-    // iconv は入力バッファを書き換えないのでキャストで const を外してよい
-    if (iconv(cd, (char **)&in, &inlen_left, &out, &outlen_left) == (size_t)-1) {
-        return -1;
-    }
-
-    return utf8_buf_len - outlen_left;
-}
-#endif
-
-/*!
- * @brief 文字コードがUTF-8の文字列をシステムの文字コードに変換する
- * @param str 変換するUTF-8の文字列
- * @param sys_str_buffer 変換したシステムの文字コードの文字列を格納するバッファへのポインタ
- * @param sys_str_buflen 変換したシステムの文字コードの文字列を格納するバッファの長さ
- * @return 変換に成功した場合TRUE、失敗した場合FALSEを返す
- */
-static bool utf8_to_sys(std::string_view str, char *sys_str_buffer, size_t sys_str_buflen)
-{
-    std::string utf8_str(str);
-
-#if defined(EUC)
-
-    /* length() + 1 を渡して文字列終端('\0')を含めて変換する */
-    return utf8_to_euc(utf8_str.data(), utf8_str.length() + 1, sys_str_buffer, sys_str_buflen) >= 0;
-
-#elif defined(SJIS) && defined(WINDOWS)
-
-    int input_len = utf8_str.length() + 1; /* include termination character */
-
-    std::vector<WCHAR> utf16buf(input_len);
-
-    /* UTF-8 -> UTF-16 */
-    if (MultiByteToWideChar(CP_UTF8, 0, utf8_str.data(), input_len, utf16buf.data(), input_len) == 0) {
-        return false;
-    }
-
-    /* UTF-8 -> SJIS(CP932) */
-    if (WideCharToMultiByte(932, 0, utf16buf.data(), -1, sys_str_buffer, sys_str_buflen, nullptr, nullptr) == 0) {
-        return false;
-    }
-
-    return true;
-
-#else
-    return false;
-#endif
-}
-
-/*!
- * @brief システムの文字コードからUTF-8に変換する
- * @param str システムの文字コードの文字列
- * @return UTF-8に変換した文字列
- *         変換に失敗した場合はtl::nullopt
- */
-tl::optional<std::string> sys_to_utf8(std::string_view str)
-{
-#if defined(EUC)
-    std::string utf8str(str.length() * 2 + 1, '\0');
-    const auto len = euc_to_utf8(str.data(), str.length(), utf8str.data(), utf8str.size());
-
-    return (len >= 0) ? tl::make_optional(std::move(utf8str.erase(len))) : tl::nullopt;
-#elif defined(SJIS) && defined(WINDOWS)
-    /* SJIS(CP932) -> UTF-16 */
-    std::vector<WCHAR> utf16buf(str.length());
-    const auto utf16_len = MultiByteToWideChar(932, 0, str.data(), str.size(), utf16buf.data(), utf16buf.size());
-    if (utf16_len == 0) {
-        return tl::nullopt;
-    }
-
-    /* UTF-16 -> UTF-8 */
-    std::vector<char> utf8buf(str.length() * 2 + 1);
-    const auto utf8_len = WideCharToMultiByte(CP_UTF8, 0, utf16buf.data(), utf16_len, utf8buf.data(), utf8buf.size(), nullptr, nullptr);
-    if (utf8_len == 0) {
-        return tl::nullopt;
-    }
-
-    return tl::make_optional<std::string>(utf8buf.data(), utf8_len);
-#else
-    return tl::nullopt;
-#endif
-}
-
-/*!
- * @brief UTF-8からシステムの文字コードに変換する
- *
- * @param str UTF-8の文字列
- * @return システムの文字コードに変換した文字列
- *         変換に失敗した場合はtl::nullopt
- */
-tl::optional<std::string> utf8_to_sys(std::string_view str)
-{
-    // UTF-8 -> SJIS or EUC でバイト長が増えることはないので、終端文字分を含めて元の文字列と同じ長さを確保しておけばよい
-    std::vector<char> sys_str_buf(str.length() + 1);
-
-    if (!utf8_to_sys(str, sys_str_buf.data(), sys_str_buf.size())) {
-        return tl::nullopt;
-    }
-
-    return tl::make_optional<std::string>(sys_str_buf.data());
-}
-
-/*!
- * @brief 受け取った文字列の文字コードを推定し、システムの文字コードへ変換する
- * @param strbuf 変換する文字列を格納したバッファへのポインタ。
- *               バッファは変換した文字列で上書きされる。
- *               UTF-8からSJISもしくはEUCへの変換を想定しているのでバッファの長さが足りなくなることはない。
- * @param buflen バッファの長さ。
- * @return 変換後の文字列の長さ（終端文字は含まない）
- */
-size_t guess_convert_to_system_encoding(char *strbuf, int buflen)
-{
-    if (is_ascii_str(strbuf)) {
-        return std::string_view(strbuf).length();
-    }
-
-    if (is_utf8_str(strbuf)) {
-        const auto sys_str = utf8_to_sys(strbuf);
-        if (!sys_str || std::ssize(*sys_str) >= buflen) {
-            msg_print("警告:文字コードの変換に失敗しました");
-            msg_erase();
-            return std::string_view(strbuf).length();
-        }
-
-        std::copy(sys_str->begin(), sys_str->end(), strbuf);
-        strbuf[sys_str->length()] = '\0';
-        return sys_str->length();
-    }
-
-    return std::string_view(strbuf).length();
 }
 
 /*!
