+++ src/main-x11.cpp
@@ -94,8 +94,8 @@
 #include "game-option/runtime-arguments.h"
 #include "game-option/special-options.h"
 #include "io/files-util.h"
+#include "locale/character-encoding.h"
 #include "locale/japanese.h"
-#include "locale/utf-8.h"
 #include "main-unix/unix-music.h"
 #include "main-unix/unix-sound.h"
 #include "main-unix/x11-type-string.h"
