+++ src/main/info-initializer.cpp
@@ -22,6 +22,7 @@
 #include "info-reader/vault-reader.h"
 #include "io/files-util.h"
 #include "io/uid-checker.h"
+#include "locale/character-encoding.h"
 #include "main/init-error-messages-table.h"
 #include "object-enchant/object-ego.h"
 #include "player-info/class-info.h"
